fix incorrect use of joi validate function
The `schema.validate` function from `joi` will return an object that has the properties `value`, `error`, `then` and `catch`. The result object is both a synchronous result with value/error and it also implements then/catch so that it fulfils the promise interface and can be awaited. If awaited, it will resolve with the value and reject with the error.

Although there is no need for this function to be asynchronous, I didn’t want to change the return type.